Implement a hack to be able to link *some* CPS-code with non-CPS modules.
Before that patch, the translation assumed that all values coming from
external modules were also in CPS form - which required to write by
hand wrappers for eg. stdlib functions.

With the current patch, the outside world is assumed to be in direct
style (which is the case for most module eg. the stdlib), except for
module names specified on the command line using the "-dcpsmodule"
option. (ex: "-dcpsmodule Foo -dcpsmodule Bar")

Then, faced to a value from an external module, the translation
recovers its type from the interface, expands it and tries to guess
the arity of the value, in order to generate a wrapper. This works for
simple cases, but NOT for e.g. functions of Printf/Scanf - in which
case it will wrongly assume they always are of arity 1 (as
eg. Printf.printf has type "('a, out_channel, unit) format -> 'a")).
(failure means that the program will segfault if Printf.printf is
applied to more than one argument)